[fix] Critical bugs: AudioCapture lifetime, shard overflow, shutdown race + build fixes
## Critical bug fixes (eng review + outside voice)

- AudioCapture lifetime: cpal Stream was dropped immediately after
  spawn_audio_pipeline() returned, silently killing audio capture.
  Fix: dedicated thread holds capture alive until cancel token fires.

- FVP header shard overflow: shard_index/shard_count were u8 (max 255).
  Large IDR keyframes at high bitrate could exceed 256 shards, causing
  silent data corruption. Fix: u8→u16 (header 8B→10B) with runtime
  guards (MAX_SHARDS=4096) on all 3 decode paths:
    1. Rust RtpDepacketizer::feed()
    2. Rust decode_packets_to_frame()
    3. C++ openxr_app.cpp FVP parser

- fvp_shutdown race: read lock for shutdown() then write lock for drop
  had a TOCTOU window. Fix: single write lock with guard.take().

## Build fixes

- build.bat: auto git submodule init (fixes openvr_driver.h not found)
- build.bat: proper ERRORLEVEL check for cmake configure + build
- CMakeLists.txt: /utf-8 flag for MSVC (fixes C4819 Unicode warnings)

## Rust warning cleanup (7 warnings → 0)

- Remove unused imports: FoveationLevel, Duration, Instant
- Remove unused variables: frame_size, controllers_clone
- Prefix unused param: _tracking in run_streaming
- Suppress dead_code: runtime/config fields in StreamingEngine
- Remove unused max_payload in test

## Other

- TODOS.md: add HeartbeatMonitor + ReconnectHandler integration task

+++ rust/common/src/protocol.rs
@@ -11,13 +11,14 @@ pub struct RtpHeader {
     pub ssrc: u32,
 }
 
-/// Focus Vision PCVR custom header (8 bytes)
+/// Focus Vision PCVR custom header (10 bytes)
+/// shard_index/count are u16 to support large IDR keyframes (>256 shards at 80Mbps).
 #[repr(C, packed)]
 #[derive(Debug, Clone, Copy)]
 pub struct FvpHeader {
     pub frame_index: u32,
-    pub shard_index: u8,
-    pub shard_count: u8,
+    pub shard_index: u16,
+    pub shard_count: u16,
     pub flags: u16,
 }
